pyproject.toml
[tool.poetry]
name = "dados-do-excel-para-o-banco"
version = "0.1.0"
description = ""
authors = ["Vinicius da Fonseca <viniciusdafonseca2000@gmail.com>"]
readme = "README.md"
packages = [{include = "dados_do_excel_para_o_banco"}]

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^1.5.1"
openpyxl = "^3.0.10"
tortoise-orm = {extras = ["asyncmy"], version = "^0.19.2"}
pydantic = "^1.10.2"
loguru = "^0.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
